+++ common/src/futures.rs
@@ -32,3 +32,27 @@ impl Executor for cpupool::CpuPool {
         self.execute(f).unwrap();
     }
 }
+
+/// Future trait with extra helper methods.
+pub trait FutureExt: Future {
+    #[cfg(target_env = "sgx")]
+    fn wait(self) -> Result<Self::Item, Self::Error>
+    where
+        Self: Sized;
+}
+
+impl<F: Future> FutureExt for F {
+    #[cfg(target_env = "sgx")]
+    fn wait(mut self) -> Result<Self::Item, Self::Error>
+    where
+        Self: Sized,
+    {
+        // Ekiden SGX enclaves are currently single-threaded and all OCALLs are blocking,
+        // so nothing should return Async::NotReady.
+        match self.poll() {
+            Ok(Async::NotReady) => panic!("futures in SGX should always block"),
+            Ok(Async::Ready(result)) => Ok(result),
+            Err(error) => Err(error),
+        }
+    }
+}