+++ consensus/api/src/lib.rs
@@ -1,8 +1,0 @@
-extern crate futures;
-extern crate grpcio;
-extern crate protobuf;
-
-mod generated;
-
-pub use generated::consensus::*;
-pub use generated::consensus_grpc::*;